requirements.txt
binance-connector==3.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pycryptodome==3.21.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
urllib3==2.2.3
websocket-client==1.8.0
